backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
langchain>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0
chromadb>=0.4.24
pypdf>=4.0.0
python-dotenv>=1.0.0
pydantic>=2.6.0
pinecone-client
langchain-pinecone
pymongo
motor